% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_portfolio.R
\name{export_portfolio}
\alias{export_portfolio}
\title{export a PACTA portfolio in a standard format}
\usage{
export_portfolio(portfolio_data, group_data, output_directory, validate = TRUE)
}
\arguments{
\item{portfolio_data}{data frame with the portfolio data}

\item{group_data}{list or data frame with the group data}

\item{output_directory}{character with the directory where the file will be}

\item{validate}{logical, should the output be validated against the schema?}
}
\value{
portfolio metadata (as nested list) for exported file, pramirly
called for side effect of writing file to disk.
}
\description{
This function takes a pacta portfolio as a data frame, and exports it to a
csv file in \code{output_directory}. optionally (default on), it can validate the
metadata that will be attached to this file export.
}
