% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_directory.R
\name{process_directory}
\alias{process_directory}
\title{re-export a directory of PACTA portfolios in a standard format}
\usage{
process_directory(
  input_directory = "/mnt/input",
  output_directory = "/mnt/output",
  validate = TRUE
)
}
\arguments{
\item{input_directory}{path to directory with input files}

\item{output_directory}{character with the directory where the file will be}

\item{validate}{logical, should the output be validated against the schema?}
}
\value{
portfolio metadata (as nested list) for exported files, primarily
called for side effect of writing files to disk.
}
\description{
This function takes a directory containing pacta portfolios as a file, and
exports them to a csv file in \code{output_directory}. optionally (default on),
it can validate the metadata that will be attached to this file export.
}
