% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_portfolio.R
\name{reexport_portfolio}
\alias{reexport_portfolio}
\title{re-export a PACTA portfolio in a standard format}
\usage{
reexport_portfolio(input_filepath, output_directory, validate = TRUE)
}
\arguments{
\item{input_filepath}{path to input file}

\item{output_directory}{character with the directory where the file will be}

\item{validate}{logical, should the output be validated against the schema?}
}
\value{
portfolio metadata (as nested list) for exported file, primarily
called for side effect of writing file to disk.
}
\description{
This function takes a pacta portfolio as a file, and exports it to a
csv file in \code{output_directory}. optionally (default on), it can validate the
metadata that will be attached to this file export.
}
